Stash broken code for Larger/Smaller/Both
It's looking like I'm going to have to check throught the files twice if I continue
to use Compare-Object.

I was bouncing around an idea where I could just look for anything in `=>`
and look at either one above or one below to just see the size.
Seems like a hack but :shrug:.